+++ src/JustSaying/AwsTools/MessageHandling/Dispatch/IMessageDispatcher.cs
@@ -1,0 +1,11 @@
+using System.Threading;
+using System.Threading.Tasks;
+using JustSaying.Messaging.Channels;
+
+namespace JustSaying.AwsTools.MessageHandling.Dispatch
+{
+    internal interface IMessageDispatcher
+    {
+        Task DispatchMessageAsync(IQueueMessageContext messageContext, CancellationToken cancellationToken);
+    }
+}
