+++ src/JustSaying/Messaging/Channels/IConsumerBus.cs
@@ -1,0 +1,11 @@
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace JustSaying.Messaging.Channels
+{
+    internal interface IConsumerBus
+    {
+        void Start(CancellationToken stoppingToken);
+        Task Completion { get; }
+    }
+}
