+++ src/JustSaying/Messaging/Channels/IMessageReceiveBuffer.cs
@@ -1,0 +1,12 @@
+using System.Threading;
+using System.Threading.Channels;
+using System.Threading.Tasks;
+
+namespace JustSaying.Messaging.Channels
+{
+    internal interface IMessageReceiveBuffer
+    {
+        Task Start(CancellationToken stoppingToken);
+        ChannelReader<IQueueMessageContext> Reader { get; }
+    }
+}
